==> scripts/3e_compare_cv_model_params_hbn.Rmd <==
---
title: "3e: Compare model parameters from selected models from CV across pipelines"
author: "Paul A. Bloom"
date: "September 10, 2019"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: paper
---

This markdown makes plots for comparing model parameter estimates from the selected models from previous cross-validation across imputation pipelines. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE)
```

## Load pacakages/Set themes
```{r}
library(tidyverse)
library(ggplot2)
library(rstanarm)
library(gridExtra)

# Plot theme
mytheme = theme_bw() 
theme_set(mytheme)

# source helper functions
source('helperFunctions.R')

# color palette
myPal3 = c('#2b8cbe', '#88419d','#cb181d')
myPal3_reorder = c('#88419d', '#2b8cbe','#cb181d')
myPal4 = c('#2b8cbe', '#88419d', '#238b45','#cb181d')
myPal4_reorder = c('#88419d', '#2b8cbe', '#238b45','#cb181d')

# To get the same random number generator across R versions (https://blog.revolutionanalytics.com/2019/05/whats-new-in-r-360.html)
if (as.numeric(R.Version()$major) >= 3 & as.numeric(R.Version()$minor) >= 6.0){ 
  RNGkind(sample.kind="Rounding")
}
```


# Child models

## Linear
```{r}
load('../output/scaredChildModels.rda')

# Pull in stan summaries
childSumComplete = completeScaredChildLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndivComplete = completeScaredChildLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')

childSum3NN= threeNNScaredChildLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndiv3NN= threeNNScaredChildLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')

childSum9NN= NineNNScaredChildLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndiv9NN= NineNNScaredChildLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')
```

Put all child linear models together in one dataframe, then relabel terms for better visualization

```{r}
# Combine info across pipelines and rename terms
childLinear = rbind(childSum3NN, childSum9NN, childSumComplete, childIndiv9NN, childIndiv3NN,childIndivComplete) %>%
  dplyr::filter(term != '(Intercept)', term != 'sigma', term != 'mean_PPD', term != 'log-posterior') %>%
  mutate(termType = case_when(
    grepl('cbcl', term) & !grepl(':', term) ~ 'GI Sum Score',
    grepl('nausea', term) & !grepl(':', term) ~ 'Nausea',
    grepl('vomi', term) & !grepl(':', term) ~ 'Vomiting',
    grepl('consti', term) & !grepl(':', term) ~ 'Constipation',
    grepl('stom', term) & !grepl(':', term) ~ 'Stomachache/Cramps',
    grepl('cbcl', term) & !grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Sex',
    grepl('cbcl', term) & grepl('age', term) & !grepl('sex', term) ~ 'GI Sum Score*Age',
    grepl('cbcl', term) & grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Age*Sex',
    term == 'sex' ~ 'Sex',
    term == 'ageCenter' ~ 'Age',
    term == 'ageCenter:sex' ~ 'Age*Sex',
    term == 'sex:ageCenter' ~ 'Age*Sex',
    grepl('nausea', term) & !grepl('age', term) & grepl('sex', term) ~ 'Nausea*Sex',
    grepl('vomit', term) & !grepl('age', term) & grepl('sex', term) ~ 'Vomiting*Sex',
    grepl('constip', term) & !grepl('age', term) & grepl('sex', term) ~ 'Constipation*Sex',
    grepl('stom', term) & !grepl('age', term) & grepl('sex', term) ~ 'Stomachache/Cramps*Sex',
    grepl('nausea', term) & !grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age',
    grepl('vomit', term) & !grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age',
    grepl('constip', term) & !grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age',
    grepl('stom', term) & !grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age',
    grepl('nausea', term) & grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age*Sex',
    grepl('vomit', term) & grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age*Sex',
    grepl('constip', term) & grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age*Sex',
    grepl('stom', term) & grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age*Sex'
  )) 

# Order terms for plotting
childLinear$termType = ordered(childLinear$termType, 
                               rev(c('GI Sum Score', 'GI Sum Score*Age', 'GI Sum Score*Sex', "GI Sum Score*Age*Sex",
                                 'Nausea', 'Constipation', 'Vomiting', 'Stomachache/Cramps', 
                                 'Nausea*Age', 'Constipation*Age', 'Vomiting*Age', 'Stomachache/Cramps*Age',
                                 'Nausea*Sex', 'Constipation*Sex', 'Vomiting*Sex', 'Stomachache/Cramps*Sex', 
                                 'Nausea*Age*Sex', 'Constipation*Age*Sex', 'Vomiting*Age*Sex',
                                 'Stomachache/Cramps*Age*Sex', 
                                 'Age', 'Sex', 'Age*Sex')))

# Plot child linear model parameters
childLinearParamsPlot = ggplot(childLinear, aes(x = termType, y = X50., color = Pipeline)) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_point(position = position_dodge(.3)) +
  geom_errorbar(aes(ymin = X2.5., ymax =X97.5.), position = position_dodge(.3), width = .2) +
  geom_errorbar(aes(ymin = X10., ymax =X90.), position = position_dodge(.3), width = 0, lwd = 1) +
  theme(axis.text.y = element_text(size = 8),
        legend.position = 'bottom') +
  labs(x = '', y = 'Beta Estimates', title = 'SCARED-C', subtitle = 'Continuous Outcome') +
  coord_flip() +
  facet_wrap(~Model, scales = 'free_y', ncol = 2) 

ggsave(childLinearParamsPlot, file = '../plots/suppParamsChildLinear.pdf', height = 8, width = 8)
```

## Same thing with child logistic regression models

```{r}
# pull in stan summaries
childSumComplete = completeScaredChildLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndivComplete = completeScaredChildLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')

childSum3NN= threeNNScaredChildLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndiv3NN= threeNNScaredChildLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')

childSum9NN= NineNNScaredChildLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Sum')

childIndiv9NN= NineNNScaredChildLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Child', term = row.names(.), Model = 'GI Indiv. Items')
```

Aggregate across models, rename terms

```{r}
childLogistic = rbind(childSum3NN, childSum9NN, childSumComplete, childIndiv9NN, childIndiv3NN,childIndivComplete) %>%
  dplyr::filter(term != '(Intercept)', term != 'sigma', term != 'mean_PPD', term != 'log-posterior') %>%
  mutate(termType = case_when(
    grepl('cbcl', term) & !grepl(':', term) ~ 'GI Sum Score',
    grepl('nausea', term) & !grepl(':', term) ~ 'Nausea',
    grepl('vomi', term) & !grepl(':', term) ~ 'Vomiting',
    grepl('consti', term) & !grepl(':', term) ~ 'Constipation',
    grepl('stom', term) & !grepl(':', term) ~ 'Stomachache/Cramps',
    grepl('cbcl', term) & !grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Sex',
    grepl('cbcl', term) & grepl('age', term) & !grepl('sex', term) ~ 'GI Sum Score*Age',
    grepl('cbcl', term) & grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Age*Sex',
    term == 'sex' ~ 'Sex',
    term == 'ageCenter' ~ 'Age',
    term == 'ageCenter:sex' ~ 'Age*Sex',
    term == 'sex:ageCenter' ~ 'Age*Sex',
    grepl('nausea', term) & !grepl('age', term) & grepl('sex', term) ~ 'Nausea*Sex',
    grepl('vomit', term) & !grepl('age', term) & grepl('sex', term) ~ 'Vomiting*Sex',
    grepl('constip', term) & !grepl('age', term) & grepl('sex', term) ~ 'Constipation*Sex',
    grepl('stom', term) & !grepl('age', term) & grepl('sex', term) ~ 'Stomachache/Cramps*Sex',
    grepl('nausea', term) & !grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age',
    grepl('vomit', term) & !grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age',
    grepl('constip', term) & !grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age',
    grepl('stom', term) & !grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age',
    grepl('nausea', term) & grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age*Sex',
    grepl('vomit', term) & grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age*Sex',
    grepl('constip', term) & grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age*Sex',
    grepl('stom', term) & grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age*Sex'
  )) 

# Rename terms for plotting
childLogistic$termType = ordered(childLogistic$termType, 
                               rev(c('GI Sum Score', 'GI Sum Score*Age', 'GI Sum Score*Sex', "GI Sum Score*Age*Sex",
                                 'Nausea', 'Constipation', 'Vomiting', 'Stomachache/Cramps', 
                                 'Nausea*Age', 'Constipation*Age', 'Vomiting*Age', 'Stomachache/Cramps*Age',
                                 'Nausea*Sex', 'Constipation*Sex', 'Vomiting*Sex', 'Stomachache/Cramps*Sex', 
                                 'Nausea*Age*Sex', 'Constipation*Age*Sex', 'Vomiting*Age*Sex',
                                 'Stomachache/Cramps*Age*Sex', 
                                 'Age', 'Sex', 'Age*Sex')))

# Plot
childLogisticParamsPlot = ggplot(childLogistic, aes(x = termType, y = X50., color = Pipeline)) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_point(position = position_dodge(.3)) +
  geom_errorbar(aes(ymin = X2.5., ymax =X97.5.), position = position_dodge(.3), width = .2) +
  geom_errorbar(aes(ymin = X10., ymax =X90.), position = position_dodge(.3), width = 0, lwd = 1) +
  theme(axis.text.y = element_text(size = 8),
        legend.position = 'bottom') +
  labs(x = '', y = 'Beta Estimates', title = 'SCARED-C', subtitle = 'Clinical Threshold Outcome') +
  coord_flip() +
  facet_wrap(~Model, scales = 'free_y', ncol = 2) 

ggsave(childLogisticParamsPlot, file = '../plots/suppParamsChildLogistic.pdf', height = 8, width = 8)
```

# Parent models

## Linear
```{r}
load('../output/scaredParentModels.rda')

# pull stan summaries
parentSumComplete = completeScaredParentLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndivComplete = completeScaredParentLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')

parentSum3NN= threeNNScaredParentLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndiv3NN= threeNNScaredParentLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')

parentSum9NN= NineNNScaredParentLinearModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndiv9NN= NineNNScaredParentLinearModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')
```

Put all parent linear models together in one dataframe, then relabel terms for better visualization

```{r}
# Combine info across pipelines and rename terms
parentLinear = rbind(parentSum3NN, parentSum9NN, parentSumComplete, parentIndiv9NN, parentIndiv3NN,parentIndivComplete) %>%
  dplyr::filter(term != '(Intercept)', term != 'sigma', term != 'mean_PPD', term != 'log-posterior') %>%
  mutate(termType = case_when(
    grepl('cbcl', term) & !grepl(':', term) ~ 'GI Sum Score',
    grepl('nausea', term) & !grepl(':', term) ~ 'Nausea',
    grepl('vomi', term) & !grepl(':', term) ~ 'Vomiting',
    grepl('consti', term) & !grepl(':', term) ~ 'Constipation',
    grepl('stom', term) & !grepl(':', term) ~ 'Stomachache/Cramps',
    grepl('cbcl', term) & !grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Sex',
    grepl('cbcl', term) & grepl('age', term) & !grepl('sex', term) ~ 'GI Sum Score*Age',
    grepl('cbcl', term) & grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Age*Sex',
    term == 'sex' ~ 'Sex',
    term == 'ageCenter' ~ 'Age',
    term == 'ageCenter:sex' ~ 'Age*Sex',
    term == 'sex:ageCenter' ~ 'Age*Sex',
    grepl('nausea', term) & !grepl('age', term) & grepl('sex', term) ~ 'Nausea*Sex',
    grepl('vomit', term) & !grepl('age', term) & grepl('sex', term) ~ 'Vomiting*Sex',
    grepl('constip', term) & !grepl('age', term) & grepl('sex', term) ~ 'Constipation*Sex',
    grepl('stom', term) & !grepl('age', term) & grepl('sex', term) ~ 'Stomachache/Cramps*Sex',
    grepl('nausea', term) & !grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age',
    grepl('vomit', term) & !grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age',
    grepl('constip', term) & !grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age',
    grepl('stom', term) & !grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age',
    grepl('nausea', term) & grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age*Sex',
    grepl('vomit', term) & grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age*Sex',
    grepl('constip', term) & grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age*Sex',
    grepl('stom', term) & grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age*Sex'
  )) 

# Rename terms for plotting
parentLinear$termType = ordered(parentLinear$termType, 
                               rev(c('GI Sum Score', 'GI Sum Score*Age', 'GI Sum Score*Sex', "GI Sum Score*Age*Sex",
                                 'Nausea', 'Constipation', 'Vomiting', 'Stomachache/Cramps', 
                                 'Nausea*Age', 'Constipation*Age', 'Vomiting*Age', 'Stomachache/Cramps*Age',
                                 'Nausea*Sex', 'Constipation*Sex', 'Vomiting*Sex', 'Stomachache/Cramps*Sex', 
                                 'Nausea*Age*Sex', 'Constipation*Age*Sex', 'Vomiting*Age*Sex',
                                 'Stomachache/Cramps*Age*Sex', 
                                 'Age', 'Sex', 'Age*Sex')))

# Plot
parentLinearParamsPlot = ggplot(parentLinear, aes(x = termType, y = X50., color = Pipeline)) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_point(position = position_dodge(.3)) +
  geom_errorbar(aes(ymin = X2.5., ymax =X97.5.), position = position_dodge(.3), width = .2) +
  geom_errorbar(aes(ymin = X10., ymax =X90.), position = position_dodge(.3), width = 0, lwd = 1) +
  theme(axis.text.y = element_text(size = 8),
        legend.position = 'bottom') +
  labs(x = '', y = 'Beta Estimates', title = 'SCARED-P', subtitle = 'Continuous Outcome') +
  coord_flip() +
  facet_wrap(~Model, scales = 'free_y', ncol = 2) 

ggsave(parentLinearParamsPlot, file = '../plots/suppParamsParentLinear.pdf', height = 8, width = 8)
```

## Same thing with Parent Logistic
```{r}
# pull stan summaries
parentSumComplete = completeScaredParentLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndivComplete = completeScaredParentLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')

parentSum3NN= threeNNScaredParentLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndiv3NN= threeNNScaredParentLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')

parentSum9NN= NineNNScaredParentLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Sum')

parentIndiv9NN= NineNNScaredParentLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Parent', term = row.names(.), Model = 'GI Indiv. Items')
```

rename terms

```{r}
parentLogistic = rbind(parentSum3NN, parentSum9NN, parentSumComplete, parentIndiv9NN, parentIndiv3NN,parentIndivComplete) %>%
  dplyr::filter(term != '(Intercept)', term != 'sigma', term != 'mean_PPD', term != 'log-posterior') %>%
  mutate(termType = case_when(
    grepl('cbcl', term) & !grepl(':', term) ~ 'GI Sum Score',
    grepl('nausea', term) & !grepl(':', term) ~ 'Nausea',
    grepl('vomi', term) & !grepl(':', term) ~ 'Vomiting',
    grepl('consti', term) & !grepl(':', term) ~ 'Constipation',
    grepl('stom', term) & !grepl(':', term) ~ 'Stomachache/Cramps',
    grepl('cbcl', term) & !grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Sex',
    grepl('cbcl', term) & grepl('age', term) & !grepl('sex', term) ~ 'GI Sum Score*Age',
    grepl('cbcl', term) & grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Age*Sex',
    term == 'sex' ~ 'Sex',
    term == 'ageCenter' ~ 'Age',
    term == 'ageCenter:sex' ~ 'Age*Sex',
    term == 'sex:ageCenter' ~ 'Age*Sex',
    grepl('nausea', term) & !grepl('age', term) & grepl('sex', term) ~ 'Nausea*Sex',
    grepl('vomit', term) & !grepl('age', term) & grepl('sex', term) ~ 'Vomiting*Sex',
    grepl('constip', term) & !grepl('age', term) & grepl('sex', term) ~ 'Constipation*Sex',
    grepl('stom', term) & !grepl('age', term) & grepl('sex', term) ~ 'Stomachache/Cramps*Sex',
    grepl('nausea', term) & !grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age',
    grepl('vomit', term) & !grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age',
    grepl('constip', term) & !grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age',
    grepl('stom', term) & !grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age',
    grepl('nausea', term) & grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age*Sex',
    grepl('vomit', term) & grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age*Sex',
    grepl('constip', term) & grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age*Sex',
    grepl('stom', term) & grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age*Sex'
  )) 

# reorder terms
parentLogistic$termType = ordered(parentLogistic$termType, 
                               rev(c('GI Sum Score', 'GI Sum Score*Age', 'GI Sum Score*Sex', "GI Sum Score*Age*Sex",
                                 'Nausea', 'Constipation', 'Vomiting', 'Stomachache/Cramps', 
                                 'Nausea*Age', 'Constipation*Age', 'Vomiting*Age', 'Stomachache/Cramps*Age',
                                 'Nausea*Sex', 'Constipation*Sex', 'Vomiting*Sex', 'Stomachache/Cramps*Sex', 
                                 'Nausea*Age*Sex', 'Constipation*Age*Sex', 'Vomiting*Age*Sex',
                                 'Stomachache/Cramps*Age*Sex', 
                                 'Age', 'Sex', 'Age*Sex')))

# plot
parentLogisticParamsPlot = ggplot(parentLogistic, aes(x = termType, y = X50., color = Pipeline)) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_point(position = position_dodge(.3)) +
  geom_errorbar(aes(ymin = X2.5., ymax =X97.5.), position = position_dodge(.3), width = .2) +
  geom_errorbar(aes(ymin = X10., ymax =X90.), position = position_dodge(.3), width = 0, lwd = 1) +
  theme(axis.text.y = element_text(size = 8),
        legend.position = 'bottom') +
  labs(x = '', y = 'Beta Estimates', title = 'SCARED-P', subtitle = 'Clinical Threshold Outcome') +
  coord_flip() +
  facet_wrap(~Model, scales = 'free_y', ncol = 2) 

ggsave(parentLogisticParamsPlot, file = '../plots/suppParamsParentLogistic.pdf', height = 8, width = 8)
```


# KSADS logistic model params

```{r}
load('../output/KsadsModels.rda')

# pull stan summaries
KsadsSumComplete = completeKsadsLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Ksads', term = row.names(.), Model = 'GI Sum')

KsadsIndivComplete = completeKsadsLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = 'Compete Cases', Reporter = 'Ksads', term = row.names(.), Model = 'GI Indiv. Items')

KsadsSum3NN= threeNNKsadsLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Ksads', term = row.names(.), Model = 'GI Sum')

KsadsIndiv3NN= threeNNKsadsLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '3NN Imputation', Reporter = 'Ksads', term = row.names(.), Model = 'GI Indiv. Items')

KsadsSum9NN= NineNNKsadsLogisticModFrame$modObject[1][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Ksads', term = row.names(.), Model = 'GI Sum')

KsadsIndiv9NN= NineNNKsadsLogisticModFrame$modObject[2][[1]]$stan_summary %>%
  data.frame(.) %>%
  mutate(Pipeline = '9NN Imputation', Reporter = 'Ksads', term = row.names(.), Model = 'GI Indiv. Items')
```

Put all KSADS logistic models together in one dataframe, then relabel terms for better visualization


```{r}
KsadsLogistic = rbind(KsadsSum3NN, KsadsSum9NN, KsadsSumComplete, KsadsIndiv9NN, KsadsIndiv3NN,KsadsIndivComplete) %>%
  dplyr::filter(term != '(Intercept)', term != 'sigma', term != 'mean_PPD', term != 'log-posterior') %>%
  mutate(termType = case_when(
    grepl('cbcl', term) & !grepl(':', term) ~ 'GI Sum Score',
    grepl('nausea', term) & !grepl(':', term) ~ 'Nausea',
    grepl('vomi', term) & !grepl(':', term) ~ 'Vomiting',
    grepl('consti', term) & !grepl(':', term) ~ 'Constipation',
    grepl('stom', term) & !grepl(':', term) ~ 'Stomachache/Cramps',
    grepl('cbcl', term) & !grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Sex',
    grepl('cbcl', term) & grepl('age', term) & !grepl('sex', term) ~ 'GI Sum Score*Age',
    grepl('cbcl', term) & grepl('age', term) & grepl('sex', term) ~ 'GI Sum Score*Age*Sex',
    term == 'sex' ~ 'Sex',
    term == 'ageCenter' ~ 'Age',
    term == 'ageCenter:sex' ~ 'Age*Sex',
    term == 'sex:ageCenter' ~ 'Age*Sex',
    grepl('nausea', term) & !grepl('age', term) & grepl('sex', term) ~ 'Nausea*Sex',
    grepl('vomit', term) & !grepl('age', term) & grepl('sex', term) ~ 'Vomiting*Sex',
    grepl('constip', term) & !grepl('age', term) & grepl('sex', term) ~ 'Constipation*Sex',
    grepl('stom', term) & !grepl('age', term) & grepl('sex', term) ~ 'Stomachache/Cramps*Sex',
    grepl('nausea', term) & !grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age',
    grepl('vomit', term) & !grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age',
    grepl('constip', term) & !grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age',
    grepl('stom', term) & !grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age',
    grepl('nausea', term) & grepl('sex', term) & grepl('age', term) ~ 'Nausea*Age*Sex',
    grepl('vomit', term) & grepl('sex', term) & grepl('age', term) ~ 'Vomiting*Age*Sex',
    grepl('constip', term) & grepl('sex', term) & grepl('age', term) ~ 'Constipation*Age*Sex',
    grepl('stom', term) & grepl('sex', term) & grepl('age', term) ~ 'Stomachache/Cramps*Age*Sex'
  )) 

# reorder terms
KsadsLogistic$termType = ordered(KsadsLogistic$termType, 
                               rev(c('GI Sum Score', 'GI Sum Score*Age', 'GI Sum Score*Sex', "GI Sum Score*Age*Sex",
                                 'Nausea', 'Constipation', 'Vomiting', 'Stomachache/Cramps', 
                                 'Nausea*Age', 'Constipation*Age', 'Vomiting*Age', 'Stomachache/Cramps*Age',
                                 'Nausea*Sex', 'Constipation*Sex', 'Vomiting*Sex', 'Stomachache/Cramps*Sex', 
                                 'Nausea*Age*Sex', 'Constipation*Age*Sex', 'Vomiting*Age*Sex',
                                 'Stomachache/Cramps*Age*Sex', 
                                 'Age', 'Sex', 'Age*Sex')))
# plot
KsadsLogisticParamsPlot = ggplot(KsadsLogistic, aes(x = termType, y = X50., color = Pipeline)) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_point(position = position_dodge(.3)) +
  geom_errorbar(aes(ymin = X2.5., ymax =X97.5.), position = position_dodge(.3), width = .2) +
  geom_errorbar(aes(ymin = X10., ymax =X90.), position = position_dodge(.3), width = 0, lwd = 1) +
  theme(axis.text.y = element_text(size = 8),
        legend.position = 'bottom') +
  labs(x = '', y = 'Beta Estimates', title = 'KSADS', subtitle = 'Clinician Consensus Diagnosis Outcome') +
  coord_flip() +
  facet_wrap(~Model, scales = 'free_y', ncol = 2) 

ggsave(KsadsLogisticParamsPlot, file = '../plots/suppParamsKsadsLogistic.pdf', height = 8, width = 8)
```

# Session Info
```{r}
sessionInfo()
```
